+++ aiotk/__init__.py
@@ -4,10 +4,12 @@
 from .mempipe import mempipe
 from .monkey import monkey_patch
 from .posix import UnixSocketServer
+from .testing import mock_subprocess
 
 
 __all__ = [
     'mempipe',
+    'mock_subprocess',
     'monkey_patch',
     'UnixSocketServer',
 ]